+++ scripts/macros.js
@@ -42,3 +42,4 @@ export {charmMonster} from './macros/2024/spells/charmMonster.js';
 export {charmPerson} from './macros/2024/spells/charmPerson.js';
 export {chillTouch} from './macros/2024/spells/chillTouch.js';
 export {chromaticOrb} from './macros/2024/spells/chromaticOrb.js';
+export {circleOfDeath} from './macros/2024/spells/circleOfDeath.js';