+++ scripts/macros/2024/spells/circleOfDeath.js
@@ -1,0 +1,5 @@
+export let circleOfDeath = {
+    name: 'Circle of Death',
+    version: '1.1.21',
+    rules: 'modern'
+};